Design-the-site-only-on-the-client-side
Design-the-site-only-on-the-client-side
You belong to a place that knows the value of your activities

Two years ago I did this design without a framework. I did not know that the framework should be used. :) :) :)

Tip 1: The home page of this site is asli.html

Tip 2: This design is in some cases difficult. For example, repeating too much of one part of the code on all pages. Or the wallpaper lacks the blur function.